- New file naming convention:  - Pages are named in all lower case kebab casing (`my-page`). This means the folder that encompasses a page(aka a route) has to be in lower case kebab casing  - Component file names must end with either a `.jsx` or `.tsx` and must be in PascalCasing (MyComponent.jsx or .tsx) to match with the definition of a React component  - Any other files should be named in camelCasing  - Some files have also been renamed for simplification  - General overview of files that were affected:   - `aboutUs` -> `about-us`   - `faqpage` -> `faqs`   - `FAQ` -> `faqs`   - `whySponsor` -> `why-sponsor`   `Navbar` -> `NavBar` - Each page.tsx now has following optional files/folders in the same level as it:  - `components` folder  - `styles.css`  - Any additionally files/folders that the page needs  - This means that the `src/components` folder is now obsolete - **NOTE**: Global components that are shared by every page must go in the `src/app/globalComponents` folder. Do not confuse the `globalComponents` folder with the `components` folder. The latter folder is for components used in the home page - With the file restructuring, the NavBar component has been moved to the `src/app/globalComponents` folder
