+++ src/app/layout.jsx
@@ -1,6 +1,6 @@
 import { Inter } from "next/font/google";
 import "./globals.css";
-import Navbar from "../components/Navbar";
+import Navbar from "./globalComponents/Navbar";
 
 const inter = Inter({ subsets: ["latin"] });
