+++ tsconfig.json
@@ -8,6 +8,7 @@
         "paths": {
       "@/*": ["./src/*"]
     },
+    "forceConsistentCasingInFileNames": true,
     "allowJs": true,
     "skipLibCheck": true,
     "strict": true,
